Move files for better organization
