+++ src/appleDeviceUtils/captureDdgGlobals.js
